Refactor to [id] route folder
